+++ go.mod
@@ -4,7 +4,6 @@ go 1.12
 
 require (
 	github.com/creachadair/otp v0.4.0
-	github.com/spf13/cobra v1.7.0 // indirect
 	golang.org/x/sys v0.1.0 // indirect
 	golang.org/x/term v0.1.0
 )
